pyproject.toml
[tool.poetry]
name = "sdemikhov_gendiff"
version = "0.4.0"
description = "Gendiff is a CLI tool to find the difference between two JSON/YAML files"
authors = ["Sergey Demikhov <demikhov.sergey@gmail.com>"]

packages = [
  { include = "gendiff" },
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Operating System :: POSIX :: Linux",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Games/Entertainment",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3.7"
]

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
flake8 = "^3.7.9"
flake8-docstrings = "^1.5.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
